replace double with long double

+++ geom/area.hpp
@@ -7,12 +7,12 @@
 
 namespace cp {
 
-template <class T> double area(const std::vector<basic_point<T>>& ps) {
+template <class T> long double area(const std::vector<basic_point<T>>& ps) {
   T ans = 0;
   for (int i = 0; i < int(ps.size()); ++i) {
     ans += cross(ps[i], ps[(i + 1) % ps.size()]);
   }
-  return std::abs(double(ans) * 0.5);
+  return std::abs((long double)(ans) * 0.5);
 }
 
 }  // namespace cp
