+++ geom/line.hpp
@@ -14,10 +14,10 @@ template <class T> class basic_line {
       : p(_p), q(_q) {}
 };
 
-using line = basic_line<double>;
+using line = basic_line<long double>;
 using iline = basic_line<long long>;
 
-template <class T> double arg(const basic_line<T>& l) { return arg(l.q - l.p); }
+template <class T> long double arg(const basic_line<T>& l) { return arg(l.q - l.p); }
 
 template <class T>
 bool is_parallel(const basic_line<T>& l, const basic_line<T>& m) {
