+++ geom/rotate.hpp
@@ -6,8 +6,8 @@
 
 namespace cp {
 
-point rotate(const point& p, double rad) {
-  double theta = arg(p) + rad;
+point rotate(const point& p, long double rad) {
+  long double theta = arg(p) + rad;
   return norm(p) * point(std::cos(theta), std::sin(theta));
 }
